// wardn/frontend/lib/api/generated/model/mCPServerDocument.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.1.0
 */
import type { _MCPServerDocumentMeta } from './_mCPServerDocumentMeta';
import type { MCPServerDocumentIconsItem } from './mCPServerDocumentIconsItem';
import type { MCPServerDocumentPackagesItem } from './mCPServerDocumentPackagesItem';
import type { MCPServerDocumentRemotesItem } from './mCPServerDocumentRemotesItem';
import type { MCPServerDocumentRepository } from './mCPServerDocumentRepository';

export interface MCPServerDocument {
  /** @minLength 1 */
  $schema: string;
  _meta?: _MCPServerDocumentMeta;
  /** @minLength 1 */
  description: string;
  icons?: MCPServerDocumentIconsItem[];
  /**
     * @minLength 3
     * @maxLength 200
     * @pattern ^[a-zA-Z0-9.-]+/[a-zA-Z0-9._-]+$
     */
  name: string;
  packages?: MCPServerDocumentPackagesItem[];
  remotes?: MCPServerDocumentRemotesItem[];
  repository?: MCPServerDocumentRepository;
  /** @maxLength 100 */
  title?: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  version: string;
  /** @maxLength 2048 */
  websiteUrl?: string;
  [key: string]: unknown;
 }